% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cbMeanPseudoBoost.R
\name{cbMeanPseudoBoost.default}
\alias{cbMeanPseudoBoost.default}
\title{Perform smoothed stagewise pseudo-value regression for the first risk in a competing risk setting and calculate confidence bands based on pointwise confidence intervals.}
\usage{
\method{cbMeanPseudoBoost}{default}(data, xmat, times, stepno = 100,
  maxstepno = 100, nu = 0.1, cv = TRUE, multicore = FALSE,
  RepMean = 50, switch.to = 150, switch.t = NULL, seed.start = NULL,
  trace = TRUE, switches = NULL, part = 0.632, RepCb = 100,
  alpha = 0.05, cv_est = TRUE, step.point = 0.001, ...)
}
\arguments{
\item{data}{A data.frame containing observation times AND statuses}

\item{xmat}{A numeric matrix containing the covariate values for all patients.}

\item{times}{A numeric vector containing the evaluation times.}

\item{stepno}{A numeric value containing the number of boosting steps to be performed. If you use cross validation (cv=TRUE), this parameter will be ignored.}

\item{maxstepno}{A numeric value containing the maximal number of boosting steps considered during the cross validation (only used if cv=TRUE).}

\item{nu}{A numeric value between 0 and 1, the shrinkage parameter for the stagewise regression algorithm. Setting it to values such as nu=0.1 avoids overfitting in early steps.}

\item{cv}{A boolean value indicating if cross validation should be performed.}

\item{multicore}{A boolean value indication if more than one core should be used for the cross validation (for this the parallel package is needed).}

\item{RepMean}{A numeric value indicating the number of modified datasets used the estimate.}

\item{switch.to}{A numeric value indicating the number of the first risk up to which the observations can be switched. It sets the value switch.t. If switch.t is specified, switch.to will be ignored.}

\item{switch.t}{A numeric value indicating the time point up to which the observations can be switched. If this value is not NULL, switch.to will be ignored.}

\item{seed.start}{A numeric value indicating the first seed value.}

\item{trace}{A boolean value indicating if additonal information should be printed out during the process.}

\item{switches}{A numeric value indicating the number of switches to be made.}

\item{part}{A numeric value between 0 and 1 indicating the fraction which should be used in the resampling approach (=0.632=}

\item{RepCb}{A numeric value indicating the number of Resampling data sets (=100)}

\item{alpha}{A numeric value between 0 and 1 indicating the type 1 error (=0.05)}

\item{cv_est}{A boolean value indicating if cross validation should be performed for the estimates seperately (=TRUE)}

\item{trace}{A boolean value indicating if additonal information should be printed to the console (=TRUE)}
}
\value{
An object of type cbMeanPseudoBoost containing the estimates and the performed boosting step number.
}
\description{
Creates RepMean modified datasets with the help of pairwise switches of the observation times with on observation time smaller than switch.t. For these dataset the function PseudoBoost will be called seperately and the mean value is taken for each effect estimate at each considered time point.
}

