% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cbSmoothPseudoBoost.R
\name{cbSmoothPseudoBoost.default}
\alias{cbSmoothPseudoBoost.default}
\title{Perform smoothed stagewise pseudo-value regression for the first risk in a competing risk setting and calculate confidence bands based on pointwise confidence intervals.}
\usage{
\method{cbSmoothPseudoBoost}{default}(data, xmat, times, stepno = 100,
  maxstepno = 100, nu = 0.1, cv = TRUE, multicore = FALSE,
  RepSmooth = 50, smooth_para = 0.02, seed.start = NULL, trace = TRUE,
  part = 0.632, RepCb = 100, alpha = 0.05, cv_est = TRUE,
  step.point = 0.001, ...)
}
\arguments{
\item{data}{A data.frame containing observation times AND statuses}

\item{xmat}{A numeric matrix containing the covariate values for all patients.}

\item{times}{A numeric vector containing the evaluation times.}

\item{stepno}{A numeric value containing the number of boosting steps to be performed. If you use cross validation (cv=TRUE), this parameter will be ignored.}

\item{maxstepno}{A numeric value containing the maximal number of boosting steps considered during the cross validation (only used if cv=TRUE).}

\item{nu}{A numeric value between 0 and 1, the shrinkage parameter for the stagewise regression algorithm. Setting it to values such as nu=0.1 avoids overfitting in early steps.}

\item{cv}{A boolean value indicating if cross validation should be performed.}

\item{multicore}{A boolean value indication if more than one core should be used for the cross validation (for this the parallel package is needed).}

\item{RepSmooth}{A numeric value indicating the number of modified datasets used the estimate.}

\item{smooth_para}{A numeric value used for smoothing with the beta-distribution (Variance of beta-distribution).}

\item{seed.start}{A numeric value indicating the first seed value.}

\item{trace}{A boolean value indicating if additonal information should be printed out during the process.}

\item{part}{A numeric value between 0 and 1 indicating the fraction which should be used in the resampling approach (=0.632=}

\item{RepCb}{A numeric value indicating the number of Resampling data sets (=100)}

\item{alpha}{A numeric value between 0 and 1 indicating the type 1 error (=0.05)}

\item{cv_est}{A boolean value indicating if cross validation should be performed for the estimates seperately (=TRUE)}

\item{step.point}{A numeric value indicating the stepwise increase of the local significance level.}
}
\value{
An object of type cbSmoothPseudoBoost containing the estimates and the performed boosting step number.
}
\description{
Creates RepSmooth modified datasets with the help of randomly changed CIF-values and corresponding observation times. For these dataset the function PseudoBoost will be called seperately and the mean value is taken for each effect estimate at each considered time point.
}

