% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cbSmoothPseudoBoost.R
\name{plot.cbSmoothPseudoBoost}
\alias{plot.cbSmoothPseudoBoost}
\title{Plot confidence bands and estimates.}
\usage{
\method{plot}{cbSmoothPseudoBoost}(object, est, ci, alpha = 0.05,
  trans = TRUE, name = "results.pdf")
}
\arguments{
\item{object}{Object of type cbSmoothPseudoBoost}

\item{est}{A numeric vector indicating the index of the estimates to be plotted.}

\item{ci}{A numeric vector indicating the index of the estimates to be plotted with confidence bands.}

\item{trans}{A boolean vector indicating if the results should be transformed (trans=TRUE => Plot exponential of the estimates).}

\item{name}{A string value indicating the name of the resulting PDF. E.g. name="results.pdf"}
}
\value{
A PDF document with the name "name".
}
\description{
Plot confidence bands and estimates.
}

