% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ciMeanPseudoBoost.R
\name{plot.ciMeanPseudoBoost}
\alias{plot.ciMeanPseudoBoost}
\title{Plot confidence intervals and estimates.}
\usage{
\method{plot}{ciMeanPseudoBoost}(object, est, ci, alpha = 0.05,
  trans = TRUE, name = "results.pdf")
}
\arguments{
\item{object}{Object of type ciMeanPseudoBoost}

\item{est}{A numeric vector indicating the index of the estimates to be plotted.}

\item{ci}{A numeric vector indicating the index of the estimates to be plotted with confidence bands.}

\item{trans}{A boolean vector indicating if the results should be transformed (trans=TRUE => Plot exponential of the estimates).}

\item{name}{A string value indicating the name of the resulting PDF. E.g. name="results.pdf"}
}
\value{
A PDF document with the name "name".
}
\description{
Plot confidence intervals and estimates.
}

