% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/smoothPseudoBoost.R
\name{plot.smoothPseudoBoost}
\alias{plot.smoothPseudoBoost}
\title{Plot an object of class smoothPseudoBoost. The estimated effects will be plotted against the evaluation times saved in the object. Output as PDF.}
\usage{
\method{plot}{smoothPseudoBoost}(object, est, comb = TRUE, trans = TRUE,
  name = "results.pdf")
}
\arguments{
\item{object}{An object of class meanPseudoBoost created by smoothPseudoBoost().}

\item{est}{A numeric vector containing the indices of the estimated to be plotted. est=c(1) will e.g. plot the estimated intercept.}

\item{comb}{A boolean vector indicating if the results should be combined in one plot (comb=TRUE) or if each estimate should be plotted seperately (comb=FALSE).}

\item{trans}{A boolean vector indicating if the results should be transformed (trans=TRUE => Plot exponential of the estimates).}

\item{name}{A string value indicating the name of the resulting PDF. E.g. name="results.pdf"}
}
\value{
A PDF document with the name "name".
}
\description{
Plot an object of class smoothPseudoBoost. The estimated effects will be plotted against the evaluation times saved in the object. Output as PDF.
}

