% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PseudoBoost.R
\name{predict.PseudoBoost}
\alias{predict.PseudoBoost}
\title{Calculates predicted linear predictor or response (CIF) based on PseudoBoost-Object}
\usage{
\method{predict}{PseudoBoost}(object, newdata, subset = NULL,
  at.step = NULL, times = NULL, type = c("lp", "response"))
}
\arguments{
\item{object}{A PseudoBoost-object}

\item{newdata}{A n.new * p matrix with new covariate values.}

\item{subset}{An optional vecotr specifying a subset of observations to be used for evaluation}

\item{at.step}{A scalar of boosting step at which prediction is wanted.}

\item{times}{A vector with T time points where prediction is wanted.}

\item{type}{Type of prediction to be returned: "lp" gives the linear predictor, "response" the resulting CIF.}
}
\description{
Calculates predicted linear predictor or response (CIF) based on PseudoBoost-Object
}

