% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/smoothPseudoBoost.R
\name{predict.smoothPseudoBoost}
\alias{predict.smoothPseudoBoost}
\title{Calculates predicted linear predictor or response (CIF) based on smoothPseudoBoost-Object}
\usage{
\method{predict}{smoothPseudoBoost}(object, newdata, subset = NULL,
  times = NULL, type = c("lp", "response"))
}
\arguments{
\item{object}{A cbSmoothPseudoBoost-object}

\item{newdata}{A n.new * p matrix with new covariate values.}

\item{subset}{An optional vecotr specifying a subset of observations to be used for evaluation}

\item{times}{A vector with T time points where prediction is wanted.}

\item{type}{Type of prediction to be returned: "lp" gives the linear predictor, "response" the resulting CIF.}
}
\description{
Calculates predicted linear predictor or response (CIF) based on smoothPseudoBoost-Object
}

