% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cbSmoothPseudoBoost.R
\name{predictProb.cbSmoothPseudoBoost}
\alias{predictProb.cbSmoothPseudoBoost}
\title{Calculates predicted response (CIF) based on cbSmoothPseudoBoost-Object}
\usage{
predictProb.cbSmoothPseudoBoost(object, newdata, subset = NULL,
  times = NULL)
}
\arguments{
\item{object}{A cbSmoothPseudoBoost-object}

\item{newdata}{A n.new * p matrix with new covariate values.}

\item{subset}{An optional vecotr specifying a subset of observations to be used for evaluation}

\item{times}{A vector with T time points where prediction is wanted.}
}
\description{
Calculates predicted response (CIF) based on cbSmoothPseudoBoost-Object
}

