% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/smoothPseudoBoost.R
\name{predictProb.smoothPseudoBoost}
\alias{predictProb.smoothPseudoBoost}
\title{Calculates predicted response (CIF) based on smoothPseudoBoost-Object}
\usage{
predictProb.smoothPseudoBoost(object, newdata, subset = NULL, times = NULL)
}
\arguments{
\item{object}{A smoothPseudoBoost-object}

\item{newdata}{A n.new * p matrix with new covariate values.}

\item{subset}{An optional vecotr specifying a subset of observations to be used for evaluation}

\item{times}{A vector with T time points where prediction is wanted.}
}
\description{
Calculates predicted response (CIF) based on smoothPseudoBoost-Object
}

