% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PseudoBoost.R
\name{PseudoBoost.default}
\alias{PseudoBoost.default}
\title{Perform stagewise pseudo-value regression for the first risk in a competing risk setting}
\usage{
\method{PseudoBoost}{default}(ymat, xmat, stepno = 10, nu = 0.1,
  trace = FALSE)
}
\arguments{
\item{ymat}{A numeric matrix containing the pseudo-values for the CIF for the risk 1}

\item{xmat}{A numeric matrix containing the covariate values for all patients.}

\item{stepno}{A numeric value containing the number of boosting steps to be performed. If you use cross validation (cv=TRUE), this parameter will be ignored.}

\item{nu}{A numeric value between 0 and 1, the shrinkage parameter for the stagewise regression algorithm. Setting it to values such as nu=0.1 avoids overfitting in early steps.}

\item{times}{A numeric vector containing the evaluation times.}

\item{maxstepno}{A numeric value containing the maximal number of boosting steps considered during the cross validation (only used if cv=TRUE).}

\item{cv}{A boolean value indicating if cross validation should be performed.}

\item{multicore}{A boolean value indication if more than one core should be used for the cross validation (for this the parallel package is needed).}
}
\value{
An object of type PseudoBoost containing the estimates and the performed boosting step number.
}
\description{
This function performs stagewise pseudo-value regression with the help of boosting for the CIF of the interesting risk.
The ymat a numeric matrix containing the pseudo-values for the CIF for the interesting risk obtained with the help of the package "prodlim" (jackknife(prodlim(Hist(observation-times-vector,status-vector) ~ 1), times=evaluation-times-vector, cause=interesting-risk-number)).
Thereby the statuses in the data frame should be natrual numbers and coded as follows: (0=Censored, 1=Interesting risk, >1=Other risks).
Additonally you have to put in the following arguments.
}

