% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PseudoBoost}
\alias{PseudoBoost}
\title{Perform stagewise pseudo-value regression for the first risk in a competing risk setting}
\usage{
PseudoBoost(object, ...)
}
\arguments{
\item{object}{A data.frame containing observation times AND statuses or a numeric matrix containing the pseudo-values for the CIF for the risk 1}

\item{xmat}{A numeric matrix containing the covariate values for all patients.}

\item{times}{A numeric vector containing the evaluation times.}

\item{stepno}{A numeric value containing the number of boosting steps to be performed. If you use cross validation (cv=TRUE), this parameter will be ignored.}

\item{maxstepno}{A numeric value containing the maximal number of boosting steps considered during the cross validation (only used if cv=TRUE).}

\item{nu}{A numeric value between 0 and 1, the shrinkage parameter for the stagewise regression algorithm. Setting it to values such as nu=0.1 avoids overfitting in early steps.}

\item{cv}{A boolean value indicating if cross validation should be performed.}

\item{multicore}{A boolean value indication if more than one core should be used for the cross validation (for this the parallel package is needed).}
}
\value{
An object of type PseudoBoost containing the estimates and the performed boosting step number.
}
\description{
Takes in eather the data, xmat, times, stepno ....
}

